Spelling correction in OpenHAB item configuration (#50)
The sample Item configuration for OpenHAB mis-spelled the word 'guesses' as 'gueses,' which, if copied & pasted, won't parse correctly.

Also, some additional clarity for the broker title in the mqtt.cfg, as well as in the OpenHAB Item config, would be useful. I spent a long time trouble-shooting a simple problem: the Item config listed the binding as {mqtt="<[BROKER:FAMILY..., while the mqtt.cfg listed it as broker.value, and I thought the all-uppercase BROKER was the correct syntax, resulting in a non-working setup.

The broker name can be whatever you want (I've made it 'find3' here), but it needs to match letter-for-letting in the Item config.